FEAT: function send email exported into utils
